delete authFunction and authFunction.tests.tsx and merge getWeeklyPicks to apiFunctions.ts and move getWeeklyPicks test to apiFunctions.test.tsx
