+++ api/apiFunctions.ts
@@ -1,4 +1,5 @@
 import { account } from "./config"
+import { Client, Databases } from "appwrite";
 
 export async function loginAccount(user: { email: string; password: string }) {
   try {
@@ -14,4 +15,21 @@ export async function logoutAccount() {
   } catch (error) {
       console.error(error);
   }
+}
+
+export async function getWeeklyPicks() {
+    try {
+        const client = new Client();
+
+        const databases = new Databases(client);
+
+        client
+            .setEndpoint("https://cloud.appwrite.io/v1")
+            .setProject("6616ea581ef9f5521c7d");
+
+        return await databases.listDocuments("6616ea9dcb86b246a7b0", "6622aa4d2db201231e31");
+
+    } catch (err) {
+        console.error(err);
+    }
 }
